pyproject.toml
[project]
name = "home-work-30-1"
version = "0.1.0"
description = ""
authors = [
    {name = "alexander vagin",email = "sashavagin1987@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django (>=5.2.6,<6.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
